requirements.txt
beautifulsoup4==4.12.2
llama_cpp_python==0.1.83
PyYAML==6.0.1
Requests==2.31.0
rich==13.5.2
